Feat - Notice 페이지 디자인
- management/notice 삭제 -> /notice로 통합
- Notice 페이지 디자인 (카드형식 게시판)
